Move EnergyDependentMultiGaussPSF test to test_psf_analytical.py
