+++ gammapy/irf/tests/test_psf_analytical.py
@@ -54,3 +54,22 @@ class TestEnergyDependentMultiGaussPSF:
 
         # TODO: try to improve precision, so that rtol can be lowered
         assert_allclose(desired, actual.degree, rtol=0.03)
+
+
+@requires_dependency('scipy')
+@requires_data('gammapy-extra')
+def test_psf_cta_1dc():
+    filename = '$GAMMAPY_EXTRA/datasets/cta-1dc/caldb/data/cta//1dc/bcf/South_z20_50h/irf_file.fits'
+    psf_irf = EnergyDependentMultiGaussPSF.read(filename, hdu='POINT SPREAD FUNCTION')
+
+    # Check that PSF is filled with 0 for energy / offset where no PSF info is given.
+    # This is needed so that stacked PSF computation doesn't error out,
+    # trying to interpolate for observations / energies where this occurs.
+    psf = psf_irf.to_energy_dependent_table_psf('4.5 deg')
+    psf = psf.table_psf_at_energy('0.05 TeV')
+    assert_allclose(psf.evaluate(rad='0.03 deg').value, 0)
+
+    # Check that evaluation works for an energy / offset where an energy is available
+    psf = psf_irf.to_energy_dependent_table_psf('2 deg')
+    psf = psf.table_psf_at_energy('1 TeV')
+    assert_allclose(psf.containment_radius(0.68).deg, 0.053728, atol=1e-4)
