+++ gammapy/irf/tests/test_psf_table.py
@@ -9,7 +9,7 @@ from astropy.coordinates import Angle
 from ...utils.testing import requires_dependency, requires_data
 from ...datasets import gammapy_extra
 from ...datasets import FermiGalacticCenter
-from ...irf import TablePSF, EnergyDependentTablePSF, EnergyDependentMultiGaussPSF
+from ...irf import TablePSF, EnergyDependentTablePSF
 from ...image import SkyImage
 
 
@@ -115,25 +115,6 @@ def test_EnergyDependentTablePSF():
     rad_max = psf_band.containment_radius(0.99)
     actual = psf_band.kernel(ref, normalize=True, rad_max=rad_max).value.sum()
     assert_allclose(actual, desired)
-
-
-@requires_dependency('scipy')
-@requires_data('gammapy-extra')
-def test_psf_cta_1dc():
-    filename = '$GAMMAPY_EXTRA/datasets/cta-1dc/caldb/data/cta//1dc/bcf/South_z20_50h/irf_file.fits'
-    psf_irf = EnergyDependentMultiGaussPSF.read(filename, hdu='POINT SPREAD FUNCTION')
-
-    # Check that PSF is filled with 0 for energy / offset where no PSF info is given.
-    # This is needed so that stacked PSF computation doesn't error out,
-    # trying to interpolate for observations / energies where this occurs.
-    psf = psf_irf.to_energy_dependent_table_psf('4.5 deg')
-    psf = psf.table_psf_at_energy('0.05 TeV')
-    assert_allclose(psf.evaluate(rad='0.03 deg').value, 0)
-
-    # Check that evaluation works for an energy / offset where an energy is available
-    psf = psf_irf.to_energy_dependent_table_psf('2 deg')
-    psf = psf.table_psf_at_energy('1 TeV')
-    assert_allclose(psf.containment_radius(0.68).deg, 0.053728, atol=1e-4)
 
 
 @requires_data('gammapy-extra')
